Fix for bug in code used in the tutorial from the BDVAL documentation.  The ExecuteSplitsMode would fail if called directly from the command line.  Using the supplied ant scripts would not execute this section of code and worked properly. See mantis issue #1356 for more details.
The cause of the problem is due to the fact that the "stats" mode from the command line has actually changed under the hood.  The ant scripts no longer use the original mode but it is still needed to execute splits from the command line.  ExecuteSplitsMode now bypasses the command line interface and calls the original StatsMode directly.